update benchmark runner and update cluster startup to rely on shared filesystem and spit out a scheduler file

+++ tpcx_bb/cluster_configuration/cluster-startup.sh
@@ -1,38 +1,42 @@
 #IB, NVLINK, or TCP
-CLUSTER_MODE=$1
+ROLE=$1
+CLUSTER_MODE="TCP"
 USERNAME=$(whoami)
 
 MAX_SYSTEM_MEMORY=$(free -m | awk '/^Mem:/{print $2}')M
 DEVICE_MEMORY_LIMIT="25GB"
 POOL_SIZE="30GB"
 
-# Fill in your environment name and conda path on each node
-TPCX_BB_HOME="/home/$USERNAME/shared/tpcx-bb"
-CONDA_ENV_NAME="rapids-tpcx-bb"
-CONDA_ENV_PATH="/home/$USERNAME/conda/etc/profile.d/conda.sh"
+# TPCX_BB_HOME=$HOME/tpcx-bb
+# CONDA_ENV_NAME="rapids-tpcx-bb"
+# CONDA_ENV_PATH="/home/$USERNAME/conda/etc/profile.d/conda.sh"
+TPCX_BB_HOME="/home/nicholasb/dev/tpcx-bb/"
+CONDA_ENV_NAME="rapids-tpcxbb-20201028"
+CONDA_ENV_PATH="/home/nicholasb/conda/etc/profile.d/conda.sh"
 
-# TODO: Unify interface/IP setting/getting for cluster startup
-# and scheduler file
-INTERFACE="ib0"
-
-# TODO: Remove hard-coding of scheduler
-SCHEDULER=$(hostname)
-SCHEDULER_FILE=$TPCX_BB_HOME/tpcx_bb/cluster_configuration/example-cluster-scheduler.json
-LOGDIR="/tmp/tpcx-bb-dask-logs/"
-WORKER_DIR="/tmp/tpcx-bb-dask-workers/"
+# Used for writing scheduler file and logs to shared storage
+LOCAL_DIRECTORY=$HOME/dask-local-directory
+SCHEDULER_FILE=$LOCAL_DIRECTORY/scheduler.json
+LOGDIR="$LOCAL_DIRECTORY/logs"
+WORKER_DIR=/tmp/$USERNAME/tpcx-bb-dask-workers/
 
 # Purge Dask worker and log directories
-rm -rf $LOGDIR/*
-mkdir -p $LOGDIR
-rm -rf $WORKER_DIR/*
-mkdir -p $WORKER_DIR
+if [ "$ROLE" = "SCHEDULER" ]; then
+    rm -rf $LOGDIR/*
+    mkdir -p $LOGDIR
+    rm -rf $WORKER_DIR/*
+    mkdir -p $WORKER_DIR
+fi
 
 # Purge Dask config directories
 rm -rf ~/.config/dask
 
-# Activate conda environment
+# Activate conda environment and install the local library
 source $CONDA_ENV_PATH
 conda activate $CONDA_ENV_NAME
+ 
+cd $TPCX_BB_HOME/tpcx_bb
+python -m pip install .
 
 # Dask/distributed configuration
 export DASK_DISTRIBUTED__COMM__TIMEOUTS__CONNECT="100s"
@@ -42,20 +46,19 @@ export DASK_DISTRIBUTED__COMM__RETRY__DELAY__MAX="60s"
 
 
 # Setup scheduler
-if [ "$HOSTNAME" = $SCHEDULER ]; then
+if [ "$ROLE" = "SCHEDULER" ]; then
   if [ "$CLUSTER_MODE" = "NVLINK" ]; then
-     CUDA_VISIBLE_DEVICES='0' DASK_UCX__CUDA_COPY=True DASK_UCX__TCP=True DASK_UCX__NVLINK=True DASK_UCX__INFINIBAND=False DASK_UCX__RDMACM=False nohup dask-scheduler --dashboard-address 8787 --interface $INTERFACE --protocol ucx > $LOGDIR/scheduler.log 2>&1 &
+     CUDA_VISIBLE_DEVICES='0' DASK_UCX__CUDA_COPY=True DASK_UCX__TCP=True DASK_UCX__NVLINK=True DASK_UCX__INFINIBAND=False DASK_UCX__RDMACM=False nohup dask-scheduler --dashboard-address 8787 --protocol ucx --scheduler-file $SCHEDULER_FILE > $LOGDIR/scheduler.log 2>&1 &
   fi
   
   if [ "$CLUSTER_MODE" = "TCP" ]; then
-     CUDA_VISIBLE_DEVICES='0' nohup dask-scheduler --dashboard-address 8787 --interface $INTERFACE --protocol tcp > $LOGDIR/scheduler.log 2>&1 &
+     CUDA_VISIBLE_DEVICES='0' nohup dask-scheduler --dashboard-address 8787 --protocol tcp --scheduler-file $SCHEDULER_FILE > $LOGDIR/scheduler.log 2>&1 &
   fi
 fi
 
-
 # Setup workers
 if [ "$CLUSTER_MODE" = "NVLINK" ]; then
-        dask-cuda-worker --device-memory-limit $DEVICE_MEMORY_LIMIT --local-directory $WORKER_DIR  --rmm-pool-size=$POOL_SIZE --memory-limit=$MAX_SYSTEM_MEMORY --enable-tcp-over-ucx --enable-nvlink  --disable-infiniband --scheduler-file $SCHEDULER_FILE >> $LOGDIR/worker.log 2>&1 &
+    dask-cuda-worker --device-memory-limit $DEVICE_MEMORY_LIMIT --local-directory $WORKER_DIR  --rmm-pool-size=$POOL_SIZE --memory-limit=$MAX_SYSTEM_MEMORY --enable-tcp-over-ucx --enable-nvlink  --disable-infiniband --scheduler-file $SCHEDULER_FILE >> $LOGDIR/worker.log 2>&1
 fi
 
 if [ "$CLUSTER_MODE" = "TCP" ]; then
